Fix u32 truncation of shallow size for large arrays (> 4 GB)
- Change array_shallow() return type from u32 to u64 to prevent overflow
- Update ObjectMeta.shallow field from u32 to u64
- Update ObjectGraph.shallow Vec from Vec<u32> to Vec<u64>
- Add tests for large array shallow sizes exceeding u32::MAX

Large Java long[] arrays can exceed 4 GB when packed:
a 2^31 element array at 8 bytes per element = 16 GB. The u32
return type truncated these to u32::MAX, causing incorrect
retained size analysis. Now correctly handles up to u64::MAX.

+++ src/graph.rs
@@ -28,7 +28,7 @@ pub struct ObjectGraph {
     /// Object addresses, sorted ascending; `addrs[id]` is node `id`'s address.
     pub addrs: Vec<u64>,
     /// Shallow size in bytes per node.
-    pub shallow: Vec<u32>,
+    pub shallow: Vec<u64>,
     /// Interned class names; indexed by the values in [`object_class`](Self::object_class).
     pub class_names: Vec<String>,
     /// Class index per node, pointing into [`class_names`](Self::class_names).
@@ -86,9 +86,9 @@ impl ObjectGraph {
                 });
         }
 
-        let mut shallow = vec![0u32; n];
+        let mut shallow = vec![0u64; n];
         let mut object_class = vec![0u32; n];
-        let metadata: Vec<(usize, u32, u32)> = index
+        let metadata: Vec<(usize, u64, u32)> = index
             .objects
             .par_iter()
             .map(|obj| {
